Adding back the condition in the Delete Task function in app.py made redirecting after deleting tasks work again!

+++ todo_app/app.py
@@ -49,5 +49,6 @@ def view_delete_item(id):
 
 @app.route('/delete/<id>', methods=['POST'])
 def delete_task(id):
-    delete_from_tasklist(id)
+    if request.form.get('delete-task-button'):
+        delete_from_tasklist(id)
     return redirect('/')
